+++ arches/app/src/components/ControlledListManager/api.ts
@@ -21,14 +21,16 @@ export const postItemToServer = async (
     $gettext: GetText
 ) => {
     let errorText;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(
             arches.urls.controlled_list_item(item.id),
             {
                 method: "POST",
-                headers: {
-                    "X-CSRFToken": Cookies.get("csrftoken"),
-                },
+                headers: { "X-CSRFToken": token },
                 body: JSON.stringify(item),
             }
         );
@@ -54,12 +56,14 @@ export const postListToServer = async (
     $gettext: GetText
 ) => {
     let errorText;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(arches.urls.controlled_list(list.id), {
             method: "POST",
-            headers: {
-                "X-CSRFToken": Cookies.get("csrftoken"),
-            },
+            headers: { "X-CSRFToken": token },
             body: JSON.stringify(list),
         });
         if (!response.ok) {
@@ -87,12 +91,14 @@ export const upsertLabel = async (
     const url = label.id
         ? arches.urls.controlled_list_item_label(label.id)
         : arches.urls.controlled_list_item_label_add;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(url, {
             method: "POST",
-            headers: {
-                "X-CSRFToken": Cookies.get("csrftoken"),
-            },
+            headers: { "X-CSRFToken": token },
             body: JSON.stringify(label),
         });
         if (!response.ok) {
@@ -117,14 +123,16 @@ export const deleteLabel = async (
     $gettext: GetText
 ) => {
     let errorText;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(
             arches.urls.controlled_list_item_label(label.id),
             {
                 method: "DELETE",
-                headers: {
-                    "X-CSRFToken": Cookies.get("csrftoken"),
-                },
+                headers: { "X-CSRFToken": token },
             }
         );
         if (!response.ok) {
@@ -152,12 +160,14 @@ export const upsertMetadata = async (
     const url = metadata.id
         ? arches.urls.controlled_list_item_image_metadata(metadata.id)
         : arches.urls.controlled_list_item_image_metadata_add;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(url, {
             method: "POST",
-            headers: {
-                "X-CSRFToken": Cookies.get("csrftoken"),
-            },
+            headers: { "X-CSRFToken": token },
             body: JSON.stringify(metadata),
         });
         if (!response.ok) {
@@ -182,14 +192,16 @@ export const deleteMetadata = async (
     $gettext: GetText,
 ) => {
     let errorText;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(
             arches.urls.controlled_list_item_image_metadata(metadata.id),
             {
                 method: "DELETE",
-                headers: {
-                    "X-CSRFToken": Cookies.get("csrftoken"),
-                },
+                headers: { "X-CSRFToken": token },
             }
         );
         if (!response.ok) {
@@ -214,14 +226,16 @@ export const deleteImage = async(
     $gettext: GetText,
 ) => {
     let errorText;
+    const token = Cookies.get("csrftoken");
+    if (!token) {
+        return;
+    }
     try {
         const response = await fetch(
             arches.urls.controlled_list_item_image(image.id),
             {
                 method: "DELETE",
-                headers: {
-                    "X-CSRFToken": Cookies.get("csrftoken"),
-                },
+                headers: { "X-CSRFToken": token },
             }
         );
         if (!response.ok) {
